A better "dequeue" abstraction defined on reviews table
